% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_regression.R
\name{doKernelSmoothing}
\alias{doKernelSmoothing}
\title{Kernel Smoothing}
\usage{
doKernelSmoothing(gpoints, data_y, data_h, data_id, hole_id)
}
\arguments{
\item{gpoints}{grid points (locations)}

\item{data_y}{data Y to be smoothed}

\item{data_h}{bandwidths}

\item{data_id}{IDs associated with the data Y}

\item{hole_id}{IDs associated with the holes}
}
\description{
Kernel Smoothing
}
\author{
Lindesay Scott-Hayward
}
