% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_regression.R
\name{evaluate2DNonParamRegression}
\alias{evaluate2DNonParamRegression}
\title{2D Nonparametric Regression (Kernel Smoothing)}
\usage{
evaluate2DNonParamRegression(p, y, loc, Y, h_values, ind)
}
\arguments{
\item{p}{(x, y) coordinates at which to calculate the regression estimate.}

\item{y}{Original value of the parameter associated with the point {\code p}}

\item{loc}{locations (matrix of coordinates)}

\item{Y}{data to be smoothed}

\item{h_values}{bandwidths}

\item{ind}{index of points to use to calculate the regression estimate}
}
\description{
A function to calculate the regression estimate for each point p and bandwiths h1 and h2. Taken from A.W. Bowman and A. Azzalini, "Applied Smoothing Techniques for Data Analysis", Oxford, Clarendon Press, 1997, pag. 53
}
\author{
Lindesay Scott-Hayward
}
