% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_util_maths.R
\name{evaluateAutocorrelation}
\alias{evaluateAutocorrelation}
\title{Autocorrelation function for ordered data.}
\usage{
evaluateAutocorrelation(y)
}
\arguments{
\item{y}{vector of densities to find autocorrelation.}
}
\description{
This function is used to find the autocorrelation of different rows of data. These are then divided by three to find sd (sigma) bandwidth of kernel to which independence occurs
}
\author{
Lindesay Scott-Hayward
}
