% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_util_maths.R
\name{evaluateGaussianKernel}
\alias{evaluateGaussianKernel}
\title{This function evaluates a Gaussian kernel.}
\usage{
evaluateGaussianKernel(x, X, h)
}
\arguments{
\item{x}{single value or vector of point to evaluate kernel at.}

\item{X}{single value or vector of second point to evaluate kernel at.}

\item{h}{single value or vector which relates to the effective radius of the kernel.}
}
\value{
number or vector of evaluated points.
}
\description{
This function evaluates a Gaussian kernel.
}
\details{
$$\frac{1}{\sqrt(2*\pi)}\exp(\frac{-\frac{x-X}{h^2}}{2}$$
}
\examples{
x<-seq(1,10, length=100)
X<-5
h<-2

kern<-evaluateGaussianKernel(x, X, h)
plot(x, kern, type='l')

}
\author{
Lindesay Scott-Hayward
}
