% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_procedure.R
\name{evaluatePosterior}
\alias{evaluatePosterior}
\title{Function to evaluate the `posterior' density surfacer; it updates the prior surface with the information associated with the new data}
\usage{
evaluatePosterior(data_new_y, data_new_cv, data_prior_y, data_prior_cv)
}
\arguments{
\item{data_new_y}{vector of new density data}

\item{data_new_cv}{vector of coefficients of variation values for the new data (between 0 and 25)}

\item{data_prior_y}{vector of prior density data}

\item{data_prior_cv}{vector of coefficients of variation values for the prior data (between 0 and 25)}
}
\value{
Returns a list object: posterior densities, posterior CVs and posterior variances
}
\description{
Function to evaluate the `posterior' density surfacer; it updates the prior surface with the information associated with the new data
}
\examples{
x<-seq(0, 10, length=15)
y1<-rep(5, 15)
cv1<-rep(0.6, 15)

y2<- (evaluateGaussianKernel(x, 5, 1)*6) +2
cv2<-rep(0.1, 15)

posterior<-evaluatePosterior(y2, cv2, y1, cv1)

plot(x,y1, ylim=c(0,6), ylab='y', xlab='x')
points(x, y2, pch=20)
points(x, posterior$post_y, pch=20, col='red')

}
