% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{extrapolateBandwidth}
\alias{extrapolateBandwidth}
\title{Functions for bandwidth selection}
\usage{
extrapolateBandwidth(data_bw, d1, d2, buf_dat, rowauto, colauto, edge_res,
  t, b, r, l, DEBUG_MODE = FALSE)
}
\description{
Functions for bandwidth selection
}
