% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_util_maths.R
\name{fromNorm2logNorm}
\alias{fromNorm2logNorm}
\title{This function converts parameters of N(\mu, \sigma^2) to parameters of lognormal (\exp{\mu, \sigma^2})}
\usage{
fromNorm2logNorm(thetan)
}
\arguments{
\item{thetan}{list object containting the mean ({\code mu}) and the variance ({\code sigma2}) of a normal distribution.}
}
\value{
list object containing {\code muln} which is the mean of the log normal and {\code sigma2ln} which is the variance of the log normal.
}
\description{
This function converts parameters of N(\mu, \sigma^2) to parameters of lognormal (\exp{\mu, \sigma^2})
}
\author{
Lindesay Scott-Hayward
}
