% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_procedure.R
\name{runUpdateProcedure}
\alias{runUpdateProcedure}
\title{Main function to run the update and smoothing of two surfaces.}
\usage{
runUpdateProcedure(data_s, data_p, filepath = "./", DEBUG_MODE = FALSE)
}
\arguments{
\item{data_s}{The survey data (new data)}

\item{data_p}{The prior data (existing surface, e.g. RES surface)}

\item{filepath}{Pathway for the csv files to be stored. defaults to the working directory.}

\item{DEBUG_MODE}{Logical stating whether additional plots are returned.}
}
\value{
various plots, two csv files into the working directory. \code{densities_sm} gives the updated survey information in the region of the survey (plus some buffer). \code{densities_sm} gives the updated and smoothed survey embedded in the prior surface.
}
\description{
Main function to run the update and smoothing of two surfaces.
}
\examples{

data(data_s) # survey data
data(data_p) # prior data

runUpdateProcedure(data_s, data_p)

}
\author{
Lindesay Scott-Hayward

Lorenzo Milazzo
}
